Refactor : JDBC template 사용
jdbc 에서 jdbc template를 사용해서 db 연결하도록 변경
